+++ src/wordcram/WordSorterAndScaler.java
@@ -24,7 +24,7 @@ class WordSorterAndScaler {
 		
 		Word[] words = copy(rawWords);
 		Arrays.sort(words);
-		double maxWeight = words[0].weight;
+		float maxWeight = words[0].weight;
 		
 		for (Word word : words) {
 			word.weight = word.weight / maxWeight;
@@ -35,7 +35,7 @@ class WordSorterAndScaler {
 	
 	private Word[] copy(Word[] rawWords) {
 		
-		// was Arrays.copyOf(rawWords, rawWords.length);
+		// was Arrays.copyOf(rawWords, rawWords.length); - removed for Java 1.5 compatibility.
 		
 		Word[] copy = new Word[rawWords.length];
 		for(int i = 0; i < copy.length; i++) {
